PR-tweaks
