+++ plugins/ros/src/components/riScInfo/RiScSelectionCard.module.css
@@ -1,5 +1,4 @@
-.sectionHeader,
-.lockedSectionHeader {
+.sectionHeader {
   font-weight: 700;
   font-size: 0.7rem;
   text-transform: uppercase;
@@ -7,8 +6,7 @@
   background-color: inherit !important;
 }
 
-.sectionHeaderWithDivider,
-.lockedSectionHeaderWithDivider {
+.sectionHeaderWithDivider {
   border-top: 1px solid var(--ros-gray-200);
   margin-top: 4px;
   padding-top: 8px;
